Refactored project file structure
