+++ src/components/Header.jsx
@@ -1,7 +1,7 @@
 import Navbar from "react-bootstrap/Navbar";
 import Nav from "react-bootstrap/Nav";
 import { Container } from "react-bootstrap";
-import { appName } from "./constants";
+import { appName } from "../constants";
 
 export default function Header() {
   return (
